% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{asteriscos}
\alias{asteriscos}
\title{asteriscos}
\usage{
asteriscos(t)
}
\arguments{
\item{t}{The estimations data frame.}
}
\description{
This function places a numeric indicator representing the strings 0\* and 0.0\* on the estimations tabulation.
}
\details{
The function places the number 888888888 to represents a cell with the string 0\* when the absolute value is 0.
Similarly places 0.0\* on the corresponding cell when appears 0.0 on the relatives columns.
This is a useful feature in order to maintain the original format of a Excel workbook.
The key is as follow:0\* <- 888888888; 0.0\* <- 888888889
}
