% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{asteriscos_int}
\alias{asteriscos_int}
\title{asteriscos_int}
\usage{
asteriscos_int(A)
}
\arguments{
\item{A}{The confidence intervals data frame.}
}
\description{
This function places a numeric indicator representing the strings 0\*, 0.0\* and 100.0\* on the confidence intervals tabulation.
}
\details{
The function places the number 888888888 to represents a cell with the string 0\* when the absolute value is 0.
Similarly places 0.0\* on the corresponding cell when appears 0.0 on the relatives columns. The same in the 100.0 cases.
This is a useful feature in order to maintain the original format of a Excel workbook.
The key is as follow:0\* <- 888888888; 0.0\* <- 888888889; 100.0\* <- 999999999
}
