% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{da_orden}
\alias{da_orden}
\title{da_orden}
\usage{
da_orden(estimaciones, z, desagregados, abajo)
}
\arguments{
\item{estimaciones}{Data frame with the estimations.}

\item{z}{Length of the initial block.}

\item{desagregados}{Vector of labels corresponding to a elements in the first bolck.}

\item{abajo}{Correspond to the "Otros" option.}
}
\value{
Vector with the ordered row names to evaluate tabulations.
}
\description{
This function gives the order for blocks in the case of large data frames.
}
\details{
If abajo is diferent of NULL, then the function fix the "Otros" in the bottom of the blocks.
Otherwise ordered into a common way.
}
