% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{eneadora}
\alias{eneadora}
\title{eneadora}
\usage{
eneadora(lista)
}
\arguments{
\item{lista}{A list coming from the function that makes tabulations.}
}
\description{
This function places a numeric indicator representing a NA cell on a list of the typical tabulations.
}
\details{
The function places the number 888888887 to represents a cell with NA instead the string "NA" on the data frames coming as result of tabulate functions.
This is a useful feature in order to maintain the original format of a Excel workbook.
}
