% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{eneadora_cv_se}
\alias{eneadora_cv_se}
\title{eneadora_cv_se}
\usage{
eneadora_cv_se(A)
}
\arguments{
\item{A}{A data frame.}
}
\description{
This function places a numeric indicator representing a NA cell.
}
\details{
The function uses the number 888888887 to represents a cell with NA instead the string "NA".
This is a useful feature in order to maintain the original format of a Excel workbook.
}
