% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{eneadora_interv}
\alias{eneadora_interv}
\title{eneadora_interv}
\usage{
eneadora_interv(A)
}
\arguments{
\item{A}{A data frame of confidence intervals.}
}
\description{
This function places a numeric indicator representing a NA cell on the confidence intervals data frame.
}
\details{
The function uses the number 888888887 to represents a cell with NA instead the string "NA".
This is a useful feature in order to maintain the original format of a Excel workbook.
}
