% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A Hello function}
\usage{
hello(nothing = TRUE)
}
\arguments{
\item{nothing}{Do you agree dogs are the best pet? Defaults to TRUE.}
}
\description{
This function writes hello instead yourself.
}
\examples{
dogs_over_cats()
}
\keyword{Hello}
