% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{impresora}
\alias{impresora}
\title{impresora}
\usage{
impresora(wbs, lista, num_hoja, ren)
}
\arguments{
\item{wbs}{Workbook used to write the tables.}

\item{lista}{Data table to be printed.}

\item{num_hoja}{Sheet of the corresponding workbook.}
}
\description{
This function helps you to write your tabulations onto an Excel workbook.
}
\details{
The function is used to print the estimation or any other tables in the workbook.
}
\keyword{imprime}
