% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{mocha}
\alias{mocha}
\title{mocha}
\usage{
mocha(tabla, quita1 = NULL, quita2 = NULL)
}
\arguments{
\item{tabla}{Is a data frame.}

\item{quita1}{Deletes elements in the first row of a data frame.}

\item{quita2}{Deletes elements in the second column of a data frame.}
}
\value{
Data frame with some delete elements.
}
\description{
This function allows you to erase some elemnts in the first row and/or second column into a data frame.
}
\details{
Is possible combine the options and works to confident intervals.
}
