% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{n}
\alias{n}
\title{N}
\usage{
n(x)
}
\arguments{
\item{x}{Integer.}
}
\value{
vector of the form: c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", ...)
}
\description{
This function makes a vector of the length indicated but placing 0 on the left side of the first 9 numbers.
}
\details{
The function does not includes "00".
}
