% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{ordena}
\alias{ordena}
\title{ordena}
\usage{
ordena(lista, columna = 5, otros = NULL)
}
\arguments{
\item{lista}{Is a data frame. Commonly the first entry of output tabulations of this package.}

\item{columna}{Number of column to make the order.}

\item{otros}{Set the "otros" option in the bottom of the ordered vector.}
}
\value{
Vector with the row names of the data frame ordered.
}
\description{
This function allows you to order into a decreasing way a data frame by a given column. Column defautl is 5.
}
\details{
Use this function to make ordered data frames multi columns.
}
\keyword{order}
