% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{pegadora}
\alias{pegadora}
\title{pegadora}
\usage{
pegadora(lista, cuantos, prec, renglon, nombres, columna = 5)
}
\arguments{
\item{lista}{Is a data frame. Commonly the first entry of output tabulations of this package.}

\item{cuantos}{Number of data frames to be pasted.}

\item{prec}{Indicates if the data frames are estimations, coefficients of variation, etc.}

\item{renglon}{Is the number of row in each data frame to be used.}

\item{nombres}{Vector with the labels for the resulting data frame}

\item{columna}{Makes numeric the indicated column. Set default is 5.}
}
\value{
Data frame combining the correspond rows of diferent data sets.
}
\description{
This function allows you to paste the first or second, etc, row of a data frames into an easy way
}
\details{
For a suitable use you´ll should use lista = bla[[1]], and the other data frames must necessarily be called bla_1, bla_2, etc.
}
