% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{Prec}
\alias{Prec}
\title{prec}
\usage{
Prec(A, B)
}
\arguments{
\item{A}{The variable to weight.}

\item{B}{The sampling design.}
}
\description{
This function gives the statisticas presitions of a total.
}
\details{
The result is an arrange containing estimation weighted, cv(%), ci and se.
}
