% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{Prom}
\alias{Prom}
\title{Prom}
\usage{
Prom(x, y, design = asp)
}
\arguments{
\item{x}{The denominator.}

\item{y}{The denominator.}

\item{design}{Set as default asp.}
}
\description{
This function gives the statistics precisions of an arithmetical mean for a weighted variable.
}
\details{
Given by the set default for the parameter named "design" is "asp", the sampling design must be assigned with
this name.
}
