% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{Tab_2}
\alias{Tab_2}
\title{Tab_2}
\usage{
Tab_2(xx, D, z)
}
\arguments{
\item{xx}{Vector containing the names of the variables that will be used in the calculation. Usually: c("TOT", paste0("TOT_", 1:n))}

\item{D}{Variable to make a disaggregation between quotation marks.}

\item{z}{Labels for the data frame.}
}
\value{
List of four entries.
}
\description{
This function allows you to obtain a list with tabulations corresponding to estimations, coefficients of variation, confidence intervals and standard errors. It requires a data set with the appropriate expansion factor and the sampling design of the "survey" library.
}
\examples{

#Variable used in this case correspond to the ENCIG of the INEGI.
library(survey)
pregs <- c("N_TRA", "P7_3", "P7_1","FAC_TRA", "UPM_DIS", "EST_DIS","ENT")
#'#Creating total variables
t7 <- tencig[, pregs]
t7$TOT = ifelse(!t7[,pregs[1]]\%in\% '20', 1, 0)
t7$TOT_1 = ifelse(t7[,pregs[2]]\%in\%c('1','6'), 1, 0)
t7$TOT_2 = ifelse(t7[,pregs[2]]\%in\%'2', 1, 0)
t7$TOT_3 = ifelse(t7[,pregs[2]]\%in\%'3', 1, 0)
t7$TOT_4 = ifelse(t7[,pregs[2]]\%in\%'4', 1, 0)
t7$TOT_5 = ifelse(t7[,pregs[2]]\%in\%'5', 1, 0)
#Defining the sampling design using "survey".
asp <- svydesign(id = ~as.numeric(UPM_DIS), strata = ~as.numeric(EST_DIS), weight = ~FAC_TRA, data = t7)
#List with the calculations. In the first entry is the estimations resulting, CV, CI and SE, respectively.
tabulado <- Tab_2(c("TOT", paste0("TOT_",1:5)), "ENT", estados)
tabulado[[1]]
#Run this example on the console


}
\keyword{estimator,}
\keyword{survey.}
