% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{Tab_tasa}
\alias{Tab_tasa}
\title{Tab_tasa}
\usage{
Tab_tasa(xx, D, z)
}
\arguments{
\item{xx}{Vector containing the names of the variables that will be used in the calculation. Usually: c("TOT", paste0("TOT_", 1:n))}

\item{D}{Variable to make a disaggregation between quotation marks.}

\item{z}{Labels for the data frame.}
}
\value{
List of four entries.
}
\description{
This function allows you to obtain a list with rate for each 10 000 tabulations corresponding to estimations, coefficients of variation, confidence intervals and standard errors.
It requires a data set with the appropriate expansion factor and the sampling design of the "survey" library.
}
\keyword{estimator,}
\keyword{survey.}
