% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{tabuleadora}
\alias{tabuleadora}
\title{tabuleadora}
\usage{
tabuleadora(
  precision = 1:4,
  muchos,
  noms,
  tipo,
  quita1 = NULL,
  quita2 = NULL,
  con_otros = NULL,
  recorta = NULL
)
}
\arguments{
\item{precision}{Set as default 1:4.}

\item{muchos}{Number of extra tabulations besides the national to be mixed.}

\item{noms}{Vector of labels.}

\item{tipo}{Type of dis-aggregations.}

\item{quita1}{If distinct of NULL indicates NA in the first row.}

\item{quita2}{If distinct of NULL indicates MA in the second column.}

\item{con_otros}{If distinct of NULL indicates the "Otros" options buttom in the data frames.}

\item{recorta}{The number of rows in each resulting block.}
}
\description{
This function makes the arrange for many data frames mixing in a suitable way the corresponding rows.
}
\details{
You can use noms = purrr::map(.x = sector, .f = ~ c(.x, z) ), where z is a labels vector and
labels in sector will be the firs label in the resulting data frame.

When recorta!=NULL the final data frame will contain less rows in each block than the "muchos" parameter.
}
